feat: address code rabbit comments
