+++ demos/janssen-tarp/browser-extension/src/features/authentication/components/OIDCClients.tsx
@@ -5,7 +5,6 @@ import {
   ChevronLeft,
   ChevronRight,
   Zap,
-  Info,
 } from 'lucide-react';
 
 import RegisterClient from './RegisterClient';
